[LTO] Fix cloning of llvm*.used when splitting module
Refines the fix in 4df31562d67fa4f4c1a74eef731f0cbddb0ae23a to only
put globals whose defs were cloned into the split regular LTO module
on the cloned llvm*.used globals. This avoids an issue where one of the
attached values was a local that was promoted in the original module
after the module was cloned. We only need to have the values defined in
the new module on those globals.

Fixes PR49251.

Differential Revision: https://reviews.llvm.org/D97013