[OCaml] Fix buildbot failure in OCaml tests
The commit 04de6d025fffecc6a7b1ee2da72cf9db392130e3 introduced
a call to `caml_alloc_initialized_string` which seems to be
unavailable on older OCaml versions. So I'm now switching to
using `caml_alloc_string` and using a `memcpy` after that, as
is done in the rest of the file.

Buildbot failure:
https://lab.llvm.org/buildbot/#/builders/16/builds/7919