[SLP] allow matching integer min/max intrinsics as reduction ops
This is a 2nd try of:
170d7936e22
which was reverted at:
 83c707c10e6
because of crashing.

This version includes extra code and tests to avoid the known
crashing examples as discussed in PR49730.

Original commit message:
As noted in D98152, we need to patch SLP to avoid regressions when
we start canonicalizing to integer min/max intrinsics.
Most of the real work to make this possible was in:
e98388c327

Differential Revision: https://reviews.llvm.org/D98981