Reapply "[InferAttributes] Materialize all infered attributes for declaration"" and follow on patches.
This reverts commit 1590e1edba5940e72907dc658873fc9dd4a6bf1b and cb59a75810e42410fb1fa97764a2b8437a906d21.

It includes a fix for the clang test which triggered the revert.  I failed to notice this one because there was another AMDGPU llvm test with a similiar name and the exact same text in the error message.  Odd.  Since only one build bot reported the clang test, I didn't notice that one.